refactor(data-module): move parser, preprocess, features into src/data/ (Closes #57)
- Create src/data/ package with parser.py, preprocess.py, features.py
- Preserve all behavior and function signatures unchanged
- Keep src/{parser,preprocess,features}.py as re-export compatibility wrappers
- All imports from old paths continue to work seamlessly
- Full test suite passes (193 tests, parser + preprocess tests verified)

This is a pure package reorganization with zero behavior changes.
